% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dumbr.R
\name{dumbr}
\alias{dumbr}
\title{dumbr function}
\usage{
dumbr(x, n = nrow(x), slow = TRUE, sens_cols = NULL)
}
\arguments{
\item{x}{a data set of class data.frame, tibble or matrix}

\item{n}{the number of rows in the returned object}

\item{slow}{randomization method. If TRUE, the function uses a randomizing function from the random package, which takes a little more computing time. If FALSE, the function uses R´s built in random number generator, which teoretically is less safe. Defaults to TRUE.}

\item{sens_cols}{convenience function that drops columns containing sensitive values (e.g. personal identifiers, account numbers etc.). You can just as well pipe the output into select() for the same result. Input should be a character vector with relevant column names.}
}
\value{
An object of the same class as x.
}
\description{
Shuffles your data set and makes it anonymous. Values are shuffled infepencently by columns.
}
\details{
ote that the function uses the values in your data. If it contains ID variables such as
social security numbers its values will still appear in the output.
If you use a n that is higher than the number of rows in your original data,
duplicates may appear which may be relevant for ID variables.
}
\examples{
dumbr(starwars, n = 10, slow = TRUE, sens_cols = c("name", "height"))

}
