% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dumbr.R
\name{dumbr}
\alias{dumbr}
\title{dumbr function}
\usage{
dumbr(x, n = nrow(x), slow = TRUE)
}
\arguments{
\item{x}{a data set of class data.frame, tibble or matrix}

\item{n}{the number of rows in the returned object}

\item{slow}{randomization method. If TRUE, the function uses a randomizing function from the random package, which takes a little more computing time. If FALSE, the function uses R´s built in random number generator, which teoretically is less safe. Defaults to TRUE.}
}
\value{
An object of the same class as x.
}
\description{
Shuffles your data set and makes it anonymous
}
\details{
NB! Note that the function uses the values in your data. If it contains ID variables such as
social security numbers its values  will still appear in the output.
If you use a n which is higher than the number of rows in your original data, duplicates may appear which may be relevant in ID variables.
}
\examples{
dumbr(starwars, n = 10, slow = TRUE)

}
